Renamed package to org.dojo.grep
