+++ pom.xml
@@ -69,7 +69,7 @@
                             </filters>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>org.dojo.Grep</mainClass>
+                                    <mainClass>org.dojo.grep.Main</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
