+++ src/main/java/org/dojo/grep/FileWriterImpl.java
@@ -1,4 +1,4 @@
-package org.dojo;
+package org.dojo.grep;
 
 import java.util.List;
