+++ src/main/java/org/dojo/grep/FilterAndWriter.java
@@ -1,4 +1,4 @@
-package org.dojo;
+package org.dojo.grep;
 
 public interface FilterAndWriter {
     void execute();
