+++ src/main/java/org/dojo/grep/Grep.java
@@ -1,4 +1,4 @@
-package org.dojo;
+package org.dojo.grep;
 
 import picocli.CommandLine;
