+++ src/main/java/org/dojo/grep/GrepParameters.java
@@ -1,4 +1,4 @@
-package org.dojo;
+package org.dojo.grep;
 
 public record GrepParameters(String regex, String path, boolean recursive, boolean ignoreCase, boolean invertMatch) {
 }
