+++ src/main/java/org/dojo/grep/Main.java
@@ -1,4 +1,4 @@
-package org.dojo;
+package org.dojo.grep;
 
 import picocli.CommandLine;
