+++ src/main/java/org/dojo/grep/RecursiveDirectoryProcessor.java
@@ -1,4 +1,4 @@
-package org.dojo;
+package org.dojo.grep;
 
 import java.nio.file.Path;
