+++ src/main/java/org/dojo/grep/Writer.java
@@ -1,4 +1,4 @@
-package org.dojo;
+package org.dojo.grep;
 
 import java.util.List;
