+++ src/test/java/org/dojo/grep/GrepTests.java
@@ -1,4 +1,4 @@
-package org.dojo;
+package org.dojo.grep;
 
 import org.junit.jupiter.api.Test;
